+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Yii Framework 2 imagine extension Change Log
 2.1.1 under development
 -----------------------
 
-- no changes in this release.
+- Enh #22: Added method `Image::resize()` to ease resizing images to fit certain dimensions.
 
 
 2.1.0 November 3, 2016